+++ packages/backend/src/models/SpaceModel.ts
@@ -1,8 +1,14 @@
 import {
     ChartKind,
+    convertOrganizationRoleToProjectRole,
+    convertProjectRoleToSpaceRole,
+    getHighestProjectRole,
+    GroupRole,
     NotFoundError,
     OrganizationMemberRole,
+    OrganizationRole,
     ProjectMemberRole,
+    ProjectRole,
     Space,
     SpaceDashboard,
     SpaceQuery,
@@ -12,7 +18,6 @@ import {
 } from '@lightdash/common';
 import * as Sentry from '@sentry/node';
 import { Knex } from 'knex';
-import { getProjectRoleOrInheritedFromOrganization } from '../controllers/authenticationRoles';
 import {
     AnalyticsChartViewsTableName,
     AnalyticsDashboardViewsTableName,
@@ -21,6 +26,8 @@ import {
     DashboardsTableName,
     DashboardVersionsTableName,
 } from '../database/entities/dashboards';
+import { EmailTableName } from '../database/entities/emails';
+import { GroupMembershipTableName } from '../database/entities/groupMemberships';
 import { OrganizationMembershipsTableName } from '../database/entities/organizationMemberships';
 import {
     DbOrganization,
@@ -34,6 +41,7 @@ import {
     PinnedListTableName,
     PinnedSpaceTableName,
 } from '../database/entities/pinnedList';
+import { ProjectGroupAccessTableName } from '../database/entities/projectGroupAccess';
 import { ProjectMembershipsTableName } from '../database/entities/projectMemberships';
 import { DbProject, ProjectTableName } from '../database/entities/projects';
 import { SavedChartsTableName } from '../database/entities/savedCharts';
@@ -356,17 +364,7 @@ export class SpaceModel {
 
     async getSpaceAccess(spaceUuid: string): Promise<SpaceShare[]> {
         const access = await this.database
-            .table(SpaceShareTableName)
-            .leftJoin(
-                SpaceTableName,
-                `${SpaceShareTableName}.space_id`,
-                `${SpaceTableName}.space_id`,
-            )
-            .leftJoin(
-                UserTableName,
-                `${UserTableName}.user_id`,
-                `${SpaceShareTableName}.user_id`,
-            )
+            .table(SpaceTableName)
             .leftJoin(
                 ProjectTableName,
                 `${SpaceTableName}.project_id`,
@@ -374,41 +372,127 @@ export class SpaceModel {
             )
             .leftJoin(
                 OrganizationMembershipsTableName,
-                `${OrganizationMembershipsTableName}.user_id`,
-                `${UserTableName}.user_id`,
+                `${OrganizationMembershipsTableName}.organization_id`,
+                `${ProjectTableName}.organization_id`,
             )
             .leftJoin(
                 ProjectMembershipsTableName,
                 function joinProjectMembershipTable() {
                     this.on(
-                        `${UserTableName}.user_id`,
+                        `${OrganizationMembershipsTableName}.user_id`,
                         '=',
                         `${ProjectMembershipsTableName}.user_id`,
                     ).andOn(
-                        `${SpaceTableName}.project_id`,
+                        `${ProjectTableName}.project_id`,
                         '=',
                         `${ProjectMembershipsTableName}.project_id`,
                     );
                 },
+            )
+            .leftJoin(SpaceShareTableName, function joinSpaceShareTable() {
+                this.on(
+                    `${OrganizationMembershipsTableName}.user_id`,
+                    '=',
+                    `${SpaceShareTableName}.user_id`,
+                ).andOn(
+                    `${SpaceTableName}.space_id`,
+                    '=',
+                    `${SpaceShareTableName}.space_id`,
+                );
+            })
+            .leftJoin(
+                GroupMembershipTableName,
+                `${OrganizationMembershipsTableName}.user_id`,
+                `${GroupMembershipTableName}.user_id`,
+            )
+            .leftJoin(
+                ProjectGroupAccessTableName,
+                function joinProjectGroupAccessTable() {
+                    this.on(
+                        `${GroupMembershipTableName}.group_uuid`,
+                        '=',
+                        `${ProjectGroupAccessTableName}.group_uuid`,
+                    ).andOn(
+                        `${ProjectTableName}.project_uuid`,
+                        '=',
+                        `${ProjectGroupAccessTableName}.project_uuid`,
+                    );
+                },
+            )
+            .leftJoin(
+                UserTableName,
+                `${OrganizationMembershipsTableName}.user_id`,
+                `${UserTableName}.user_id`,
+            )
+            .innerJoin(
+                EmailTableName,
+                `${UserTableName}.user_id`,
+                `${EmailTableName}.user_id`,
+            )
+            .where(`${EmailTableName}.is_primary`, true)
+            .where(`${SpaceTableName}.space_uuid`, spaceUuid)
+            .where((query) => {
+                query
+                    .where((query1) => {
+                        // if space is private, only return user with direct access or admin role
+                        query1
+                            .where(`${SpaceTableName}.is_private`, true)
+                            .andWhere((query2) => {
+                                query2
+                                    .whereNotNull(
+                                        `${SpaceShareTableName}.user_id`,
+                                    )
+                                    .orWhere(
+                                        `${ProjectMembershipsTableName}.role`,
+                                        'admin',
+                                    )
+                                    .orWhere(
+                                        `${OrganizationMembershipsTableName}.role`,
+                                        'admin',
+                                    )
+                                    .orWhere(
+                                        `${ProjectGroupAccessTableName}.role`,
+                                        'admin',
+                                    );
+                            });
+                    })
+                    .orWhere(`${SpaceTableName}.is_private`, false);
+            })
+            .distinctOn(`${UserTableName}.user_uuid`)
+            .groupBy(
+                `${UserTableName}.user_id`,
+                `${UserTableName}.first_name`,
+                `${UserTableName}.last_name`,
+                `${EmailTableName}.email`,
+                `${ProjectMembershipsTableName}.role`,
+                `${OrganizationMembershipsTableName}.role`,
+                `${SpaceShareTableName}.user_id`,
             )
             .select<
                 {
                     user_uuid: string;
                     first_name: string;
                     last_name: string;
+                    email: string;
+                    user_with_direct_access: boolean;
                     project_role: ProjectMemberRole | null;
                     organization_role: OrganizationMemberRole;
+                    group_roles: (ProjectMemberRole | null)[];
                 }[]
             >([
                 `users.user_uuid`,
                 `users.first_name`,
                 `users.last_name`,
-
+                `emails.email`,
+                this.database.raw(
+                    `CASE WHEN ${SpaceShareTableName}.user_id IS NULL THEN false ELSE true end as user_with_direct_access`,
+                ),
                 `${ProjectMembershipsTableName}.role as project_role`,
                 `${OrganizationMembershipsTableName}.role as organization_role`,
-            ])
-            .distinctOn(`users.user_uuid`)
-            .where(`${SpaceTableName}.space_uuid`, spaceUuid);
+                this.database.raw(
+                    `array_agg(${ProjectGroupAccessTableName}.role) as group_roles`,
+                ),
+            ]);
 
         return access.reduce<SpaceShare[]>(
             (
@@ -417,17 +501,38 @@ export class SpaceModel {
                     user_uuid,
                     first_name,
                     last_name,
+                    email,
+                    user_with_direct_access,
                     project_role,
                     organization_role,
+                    group_roles,
                 },
             ) => {
-                const role = getProjectRoleOrInheritedFromOrganization(
-                    project_role,
-                    organization_role,
+                const inheritedOrgRole: OrganizationRole = {
+                    type: 'organization',
+                    role: convertOrganizationRoleToProjectRole(
+                        organization_role,
+                    ),
+                };
+
+                const inheritedProjectRole: ProjectRole = {
+                    type: 'project',
+                    role: project_role ?? undefined,
+                };
+
+                const inheritedGroupRoles: GroupRole[] = group_roles.map(
+                    (role) => ({ type: 'group', role: role ?? undefined }),
                 );
+
+                const highestRole = getHighestProjectRole([
+                    inheritedOrgRole,
+                    inheritedProjectRole,
+                    ...inheritedGroupRoles,
+                ]);
+
                 // exclude all users that were converted to organization members and have no space access
-                if (!role) {
-                    this.removeSpaceAccess(spaceUuid, user_uuid);
+                if (!highestRole) {
+                    this.removeSpaceAccess(spaceUuid, user_uuid); // remove access from the space if it exists
                     return acc;
                 }
                 return [
@@ -436,7 +541,11 @@ export class SpaceModel {
                         userUuid: user_uuid,
                         firstName: first_name,
                         lastName: last_name,
-                        role,
+                        email,
+                        role: convertProjectRoleToSpaceRole(highestRole.role),
+                        hasDirectAccess: !!user_with_direct_access,
+                        inheritedRole: highestRole.role,
+                        inheritedFrom: highestRole.type,
                     },
                 ];
             },