+++ packages/frontend/src/components/NavBar/index.tsx
@@ -1,4 +1,4 @@
-import { ProjectType } from '@lightdash/common';
+import { FeatureFlags, ProjectType } from '@lightdash/common';
 import {
     ActionIcon,
     Box,
@@ -13,8 +13,10 @@ import {
     Tooltip,
 } from '@mantine/core';
 import { IconInfoCircle, IconTool } from '@tabler/icons-react';
+import { useFeatureFlagEnabled } from 'posthog-js/react';
 import { FC, memo, useEffect, useMemo } from 'react';
 import { Link, useHistory, useParams } from 'react-router-dom';
+import Omnibar from '../../features/omnibar';
 import useDashboardStorage from '../../hooks/dashboard/useDashboardStorage';
 import { useActiveProjectUuid } from '../../hooks/useActiveProject';
 import { useProjects } from '../../hooks/useProjects';
@@ -113,6 +115,8 @@ const NavBar = memo(() => {
         clearDashboardStorage,
     } = useDashboardStorage();
 
+    const useOmnibar = useFeatureFlagEnabled(FeatureFlags.Omnibar);
+
     const { isFullscreen } = useApp();
 
     const dashboardInfo = getEditingDashboardInfo();
@@ -187,7 +191,11 @@ const NavBar = memo(() => {
                                 color="gray.8"
                             />
 
-                            <GlobalSearch projectUuid={activeProjectUuid} />
+                            {useOmnibar ? (
+                                <Omnibar projectUuid={activeProjectUuid} />
+                            ) : (
+                                <GlobalSearch projectUuid={activeProjectUuid} />
+                            )}
                         </>
                     ) : null}
                 </Group>