+++ packages/frontend/src/features/omnibar/types/searchItem.ts
@@ -1,0 +1,12 @@
+import { SearchResult } from '@lightdash/common';
+
+export type SearchItem = {
+    type: 'space' | 'dashboard' | 'saved_chart' | 'table' | 'field' | 'page';
+    typeLabel?: 'Table' | 'Joined table' | 'Dimension' | 'Metric';
+    title: string;
+    prefix?: string;
+    description?: string;
+    location: { pathname: string; search?: string };
+    item?: SearchResult;
+    searchRank?: number;
+};